feat: funcionamiento megustas en recetas

+++ database/migrations/2026_05_03_220510_remove_photo_from_recipes_table.php
@@ -9,10 +9,12 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-   public function up(): void
+public function up(): void
 {
     Schema::table('recipes', function (Blueprint $table) {
-        $table->dropColumn('photo');
+        if (Schema::hasColumn('recipes', 'photo')) {
+            $table->dropColumn('photo');
+        }
     });
 }
